personal_finance_tracker/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator
